Fixed some validation issues (still not good though).
Moved the picture function into draw_graph.py for easier reach and understanding of the workings.

Fixed a critical error with the graph where the max and mins for each axis was mixed causing the graph to be slighly incorrect.

Added an images folder for organisation and included a gif of the final example.