Major changes have been made in this commit.
The point of the program has now changed for any two players as long as the data is provided. The names for the players should be provided in data.txt, with line 1 being "0" and line 2 being "1". Therefore the program no longer is specifically for zeus and s1mple from Na'Vi but for any two players. I may add a feature to pull training data from HLTV.org in the future.

I also fixed a major issue where an incorrect number of points were added due to incorrect axes (again...).